requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
pytest==6.1.2
requests==2.25.0
